Write renderSprites and intToBinaryString
renderSprites take x and y coordinates and a sprite array that represents
the sprite that's going to be displayed with a grid of 1s and 0s. Each element in this array is a byte or
row of the sprite. It then parses this array rendering the rows downwards and rightwards from the origin.
For this, we have to extract the current state of where the sprite is going to be placed and perform an xor
operation between the screen rows and the sprite rows. After this we can render the sprite row by row with the
output of the xor operation.

intToBinaryString takes a number and turns it into binary in the form of a string.