Implement enough instructions to run IBM rom
app.js takes care of the logic of the emulator. All memory addresses,
registers, program counter, timers, flags are declared in the chip emulator
module pattern. As well as the emulation cycle, loading roms and the mapping
of instructions to the emulated behaviour in the browser.

This logic next to the chip display module (which is responsible of keeping the state
of the display in a buffer array for later display) are used together in the main function
to achieve the emulation of chip 8 roms. Only the IBM logo so far.